@@ -1,34 +1,51 @@
 {
-  "name": "khanamiryan/qrcode-detector-decoder",
-  "type": "library",
-  "description": "QR code decoder / reader",
-  "keywords": [
-    "qr",
-    "zxing",
-    "barcode"
-  ],
-  "homepage": "https://github.com/khanamiryan/php-qrcode-detector-decoder/",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Ashot Khanamiryan",
-      "email": "a.khanamiryan@gmail.com",
-      "homepage": "https://github.com/khanamiryan",
-      "role": "Developer"
-    }
-  ],
-  "require": {
-    "php": "^5.6|^7.0"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^5.7"
-  },
-  "autoload": {
-    "psr-4": {
-      "Zxing\\": "lib/"
-    },
-    "files": [
-      "lib/Common/customFunctions.php"
-    ]
-  }
+	"name": "chillerlan/php-qrcode-decoder",
+	"type": "library",
+	"description": "QR code decoder / reader",
+	"keywords": [
+		"qr",
+		"zxing",
+		"barcode"
+	],
+	"homepage": "https://github.com/codemasher/php-qrcode-decoder",
+	"license": "MIT",
+	"authors": [
+		{
+			"name": "Ashot Khanamiryan",
+			"email": "a.khanamiryan@gmail.com",
+			"homepage": "https://github.com/khanamiryan",
+			"role": "Developer"
+		},
+		{
+			"name": "Smiley",
+			"email": "smiley@chillerlan.net",
+			"homepage": "https://github.com/codemasher",
+			"role": "Developer"
+		},
+		{
+			"name": "Contributors",
+			"homepage":"https://github.com/chillerlan/php-qrcode/graphs/contributors"
+		}
+	],
+	"require": {
+		"php": "^7.4",
+		"ext-gd": "*",
+		"ext-iconv": "*"
+	},
+	"require-dev": {
+		"phpunit/phpunit": "^9.0"
+	},
+	"autoload": {
+		"psr-4": {
+			"Zxing\\": "lib/"
+		},
+		"files": [
+			"lib/Common/customFunctions.php"
+		]
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"Khanamiryan\\QrCodeTests\\": "tests/"
+		}
+	}
 }
