@@ -17,6 +17,7 @@
 
 namespace Zxing;
 
+use Zxing\Common\BitArray;
 use Zxing\Common\BitMatrix;
 
 /**
@@ -29,16 +30,16 @@ use Zxing\Common\BitMatrix;
  */
 abstract class Binarizer{
 
-	private $source;
+	protected LuminanceSource $source;
 
-	protected function __construct($source){
+	protected function __construct(LuminanceSource $source){
 		$this->source = $source;
 	}
 
 	/**
 	 * @return LuminanceSource
 	 */
-	public final function getLuminanceSource(){
+	final public function getLuminanceSource():LuminanceSource{
 		return $this->source;
 	}
 
@@ -50,14 +51,14 @@ abstract class Binarizer{
 	 * and passed in with each call for performance. However it is legal to keep more than one row
 	 * at a time if needed.
 	 *
-	 * @param y   The row to fetch, which must be in [0, bitmap height)
-	 * @param row An optional preallocated array. If null or too small, it will be ignored.
+	 * @param int $y   The row to fetch, which must be in [0, bitmap height)
+	 * @param \Zxing\Common\BitArray $row An optional preallocated array. If null or too small, it will be ignored.
 	 *            If used, the Binarizer will call BitArray.clear(). Always use the returned object.
 	 *
-	 * @return array The array of bits for this row (true means black).
+	 * @return BitArray The array of bits for this row (true means black).
 	 * @throws NotFoundException if row can't be binarized
 	 */
-	public abstract function getBlackRow($y, $row);
+	abstract public function getBlackRow(int $y, BitArray $row = null):BitArray;
 
 	/**
 	 * Converts a 2D array of luminance data to 1 bit data. As above, assume this method is expensive
@@ -68,24 +69,24 @@ abstract class Binarizer{
 	 * @return BitMatrix The 2D array of bits for the image (true means black).
 	 * @throws NotFoundException if image can't be binarized to make a matrix
 	 */
-	public abstract function getBlackMatrix();
+	abstract public function getBlackMatrix():BitMatrix;
 
 	/**
 	 * Creates a new object with the same type as this Binarizer implementation, but with pristine
 	 * state. This is needed because Binarizer implementations may be stateful, e.g. keeping a cache
 	 * of 1 bit data. See Effective Java for why we can't use Java's clone() method.
 	 *
-	 * @param source The LuminanceSource this Binarizer will operate on.
+	 * @param LuminanceSource source The LuminanceSource this Binarizer will operate on.
 	 *
 	 * @return Binarizer A new concrete Binarizer implementation object.
 	 */
-	public abstract function createBinarizer($source);
+	abstract public function createBinarizer(LuminanceSource $source):Binarizer;
 
-	public final function getWidth(){
+	final public function getWidth():int{
 		return $this->source->getWidth();
 	}
 
-	public final function getHeight(){
+	final public function getHeight():int{
 		return $this->source->getHeight();
 	}
 }
