@@ -18,6 +18,7 @@
 namespace Zxing;
 
 use InvalidArgumentException;
+use Zxing\Common\BitArray;
 use Zxing\Common\BitMatrix;
 
 /**
@@ -28,27 +29,24 @@ use Zxing\Common\BitMatrix;
  */
 final class BinaryBitmap{
 
-	private $binarizer;
-	private $matrix;
+	private Binarizer $binarizer;
+	private ?BitMatrix $matrix = null;
 
 	public function __construct(Binarizer $binarizer){
-		if($binarizer === null){
-			throw new InvalidArgumentException("Binarizer must be non-null.");
-		}
 		$this->binarizer = $binarizer;
 	}
 
 	/**
 	 * @return int The width of the bitmap.
 	 */
-	public function getWidth(){
+	public function getWidth():int{
 		return $this->binarizer->getWidth();
 	}
 
 	/**
 	 * @return int The height of the bitmap.
 	 */
-	public function getHeight(){
+	public function getHeight():int{
 		return $this->binarizer->getHeight();
 	}
 
@@ -57,21 +55,21 @@ final class BinaryBitmap{
 	 * cached data. Callers should assume this method is expensive and call it as seldom as possible.
 	 * This method is intended for decoding 1D barcodes and may choose to apply sharpening.
 	 *
-	 * @param y   The row to fetch, which must be in [0, bitmap height)
-	 * @param row An optional preallocated array. If null or too small, it will be ignored.
+	 * @param int $y   The row to fetch, which must be in [0, bitmap height)
+	 * @param \Zxing\Common\BitArray $row An optional preallocated array. If null or too small, it will be ignored.
 	 *            If used, the Binarizer will call BitArray.clear(). Always use the returned object.
 	 *
-	 * @return array The array of bits for this row (true means black).
+	 * @return BitArray The array of bits for this row (true means black).
 	 * @throws NotFoundException if row can't be binarized
 	 */
-	public function getBlackRow($y, $row){
+	public function getBlackRow(int $y, BitArray $row = null):BitArray{
 		return $this->binarizer->getBlackRow($y, $row);
 	}
 
 	/**
 	 * @return bool Whether this bitmap can be cropped.
 	 */
-	public function isCropSupported(){
+	public function isCropSupported():bool{
 		return $this->binarizer->getLuminanceSource()->isCropSupported();
 	}
 
@@ -79,23 +77,23 @@ final class BinaryBitmap{
 	 * Returns a new object with cropped image data. Implementations may keep a reference to the
 	 * original data rather than a copy. Only callable if isCropSupported() is true.
 	 *
-	 * @param left   The left coordinate, which must be in [0,getWidth())
-	 * @param top    The top coordinate, which must be in [0,getHeight())
-	 * @param width  The width of the rectangle to crop.
-	 * @param height The height of the rectangle to crop.
+	 * @param int $left   The left coordinate, which must be in [0,getWidth())
+	 * @param int $top    The top coordinate, which must be in [0,getHeight())
+	 * @param int $width  The width of the rectangle to crop.
+	 * @param int $height The height of the rectangle to crop.
 	 *
 	 * @return BinaryBitmap A cropped version of this object.
 	 */
-	public function crop($left, $top, $width, $height){
+	public function crop(int $left, int $top, int $width, int $height):BinaryBitmap{
 		$newSource = $this->binarizer->getLuminanceSource()->crop($left, $top, $width, $height);
 
 		return new BinaryBitmap($this->binarizer->createBinarizer($newSource));
 	}
 
 	/**
-	 * @return Whether this bitmap supports counter-clockwise rotation.
+	 * @return bool Whether this bitmap supports counter-clockwise rotation.
 	 */
-	public function isRotateSupported(){
+	public function isRotateSupported():bool{
 		return $this->binarizer->getLuminanceSource()->isRotateSupported();
 	}
 
@@ -103,9 +101,9 @@ final class BinaryBitmap{
 	 * Returns a new object with rotated image data by 90 degrees counterclockwise.
 	 * Only callable if {@link #isRotateSupported()} is true.
 	 *
-	 * @return BinaryBitmap A rotated version of this object.
+	 * @return \Zxing\BinaryBitmap A rotated version of this object.
 	 */
-	public function rotateCounterClockwise(){
+	public function rotateCounterClockwise():BinaryBitmap{
 		$newSource = $this->binarizer->getLuminanceSource()->rotateCounterClockwise();
 
 		return new BinaryBitmap($this->binarizer->createBinarizer($newSource));
@@ -115,15 +113,16 @@ final class BinaryBitmap{
 	 * Returns a new object with rotated image data by 45 degrees counterclockwise.
 	 * Only callable if {@link #isRotateSupported()} is true.
 	 *
-	 * @return BinaryBitmap A rotated version of this object.
+	 * @return \Zxing\BinaryBitmap A rotated version of this object.
 	 */
-	public function rotateCounterClockwise45(){
+	public function rotateCounterClockwise45():BinaryBitmap{
 		$newSource = $this->binarizer->getLuminanceSource()->rotateCounterClockwise45();
 
 		return new BinaryBitmap($this->binarizer->createBinarizer($newSource));
 	}
 
-	public function toString(){
+	public function toString():string{
+
 		try{
 			return $this->getBlackMatrix()->toString();
 		}
@@ -142,7 +141,7 @@ final class BinaryBitmap{
 	 * @return BitMatrix The 2D array of bits for the image (true means black).
 	 * @throws NotFoundException if image can't be binarized to make a matrix
 	 */
-	public function getBlackMatrix(){
+	public function getBlackMatrix():BitMatrix{
 		// The matrix is created on demand the first time it is requested, then cached. There are two
 		// reasons for this:
 		// 1. This work will never be done if the caller only installs 1D Reader objects, or if a
