@@ -30,10 +30,10 @@ use Exception;
  */
 abstract class LuminanceSource{
 
-	private $width;
-	private $height;
+	private int $width;
+	private int $height;
 
-	public function __construct($width, $height){
+	public function __construct(int $width, int $height){
 		$this->width  = $width;
 		$this->height = $height;
 	}
@@ -42,30 +42,30 @@ abstract class LuminanceSource{
 	 * Fetches luminance data for the underlying bitmap. Values should be fetched using:
 	 * {@code int luminance = array[y * width + x] & 0xff}
 	 *
-	 * @return A row-major 2D array of luminance values. Do not use result.length as it may be
+	 * @return array A row-major 2D array of luminance values. Do not use result.length as it may be
 	 *         larger than width * height bytes on some platforms. Do not modify the contents
 	 *         of the result.
 	 */
-	public abstract function getMatrix();
+	abstract public function getMatrix():array;
 
 	/**
-	 * @return The width of the bitmap.
+	 * @return int The width of the bitmap.
 	 */
-	public final function getWidth(){
+	final public function getWidth():int{
 		return $this->width;
 	}
 
 	/**
-	 * @return The height of the bitmap.
+	 * @return int The height of the bitmap.
 	 */
-	public final function getHeight(){
+	final public function getHeight():int{
 		return $this->height;
 	}
 
 	/**
 	 * @return bool Whether this subclass supports cropping.
 	 */
-	public function isCropSupported(){
+	public function isCropSupported():bool{
 		return false;
 	}
 
@@ -73,37 +73,37 @@ abstract class LuminanceSource{
 	 * Returns a new object with cropped image data. Implementations may keep a reference to the
 	 * original data rather than a copy. Only callable if isCropSupported() is true.
 	 *
-	 * @param left   The left coordinate, which must be in [0,getWidth())
-	 * @param top    The top coordinate, which must be in [0,getHeight())
-	 * @param width  The width of the rectangle to crop.
-	 * @param height The height of the rectangle to crop.
+	 * @param int $left   The left coordinate, which must be in [0,getWidth())
+	 * @param int $top    The top coordinate, which must be in [0,getHeight())
+	 * @param int $width  The width of the rectangle to crop.
+	 * @param int $height The height of the rectangle to crop.
 	 *
-	 * @return A cropped version of this object.
+	 * @return \Zxing\LuminanceSource A cropped version of this object.
 	 */
-	public function crop($left, $top, $width, $height){
+	public function crop(int $left, int $top, int $width, int $height):LuminanceSource{
 		throw new Exception("This luminance source does not support cropping.");
 	}
 
 	/**
-	 * @return Whether this subclass supports counter-clockwise rotation.
+	 * @return bool Whether this subclass supports counter-clockwise rotation.
 	 */
-	public function isRotateSupported(){
+	public function isRotateSupported():bool{
 		return false;
 	}
 
 	/**
-	 * @return a wrapper of this {@code LuminanceSource} which inverts the luminances it returns -- black becomes
+	 * @return InvertedLuminanceSource a wrapper of this {@code LuminanceSource} which inverts the luminances it returns -- black becomes
 	 *  white and vice versa, and each value becomes (255-value).
 	 */
 	public function invert(){
-		return new InvertedLuminanceSource($this);
+#		return new InvertedLuminanceSource($this);
 	}
 
 	/**
 	 * Returns a new object with rotated image data by 90 degrees counterclockwise.
 	 * Only callable if {@link #isRotateSupported()} is true.
 	 *
-	 * @return A rotated version of this object.
+	 * @return LuminanceSource rotated version of this object.
 	 */
 	public function rotateCounterClockwise(){
 		throw new Exception("This luminance source does not support rotation by 90 degrees.");
@@ -113,20 +113,23 @@ abstract class LuminanceSource{
 	 * Returns a new object with rotated image data by 45 degrees counterclockwise.
 	 * Only callable if {@link #isRotateSupported()} is true.
 	 *
-	 * @return A rotated version of this object.
+	 * @return LuminanceSource rotated version of this object.
 	 */
 	public function rotateCounterClockwise45(){
 		throw new Exception("This luminance source does not support rotation by 45 degrees.");
 	}
 
-	public final function toString(){
+	final public function toString():string{
 		$row    = [];
 		$result = '';
+
 		for($y = 0; $y < $this->height; $y++){
 			$row = $this->getRow($y, $row);
+
 			for($x = 0; $x < $this->width; $x++){
 				$luminance = $row[$x] & 0xFF;
 				$c         = '';
+
 				if($luminance < 0x40){
 					$c = '#';
 				}
@@ -139,9 +142,11 @@ abstract class LuminanceSource{
 				else{
 					$c = ' ';
 				}
-				$result .= ($c);
+
+				$result .= $c;
 			}
-			$result .= ('\n');
+
+			$result .= "\n";
 		}
 
 		return $result;
@@ -154,12 +159,12 @@ abstract class LuminanceSource{
 	 * to only fetch this row rather than the whole image, since no 2D Readers may be installed and
 	 * getMatrix() may never be called.
 	 *
-	 * @param $y   ; The row to fetch, which must be in [0,getHeight())
-	 * @param $row ; An optional preallocated array. If null or too small, it will be ignored.
+	 * @param int $y   ; The row to fetch, which must be in [0,getHeight())
+	 * @param array $row ; An optional preallocated array. If null or too small, it will be ignored.
 	 *             Always use the returned object, and ignore the .length of the array.
 	 *
 	 * @return array
 	 * An array containing the luminance data.
 	 */
-	public abstract function getRow($y, $row);
+	abstract public function getRow(int $y, array $row = null):array;
 }
