@@ -28,10 +28,10 @@ use InvalidArgumentException;
  */
 final class DataBlock{
 
-	private $numDataCodewords;
-	private $codewords; //byte[]
+	private int $numDataCodewords;
+	private array $codewords; //byte[]
 
-	private function __construct($numDataCodewords, $codewords){
+	private function __construct(int $numDataCodewords, array $codewords){
 		$this->numDataCodewords = $numDataCodewords;
 		$this->codewords        = $codewords;
 	}
@@ -41,20 +41,16 @@ final class DataBlock{
 	 * That is, the first byte of data block 1 to n is written, then the second bytes, and so on. This
 	 * method will separate the data into original blocks.</p>
 	 *
-	 * @param rawCodewords bytes as read directly from the QR Code
-	 * @param version      version of the QR Code
-	 * @param ecLevel      error-correction level of the QR Code
+	 * @param array                                      $rawCodewords bytes as read directly from the QR Code
+	 * @param \Zxing\Qrcode\Decoder\Version              $version      version of the QR Code
+	 * @param \Zxing\Qrcode\Decoder\ErrorCorrectionLevel $ecLevel      error-correction level of the QR Code
 	 *
-	 * @return array DataBlocks containing original bytes, "de-interleaved" from representation in the
-	 *         QR Code
+	 * @return \Zxing\Qrcode\Decoder\DataBlock[] DataBlocks containing original bytes, "de-interleaved" from
+	 *                                           representation in the QR Code
 	 */
-	public static function getDataBlocks(
-		$rawCodewords,
-		$version,
-		$ecLevel
-	){
+	public static function getDataBlocks(array $rawCodewords, Version $version, ErrorCorrectionLevel $ecLevel):array{
 
-		if(count($rawCodewords) != $version->getTotalCodewords()){
+		if(\count($rawCodewords) !== $version->getTotalCodewords()){
 			throw new InvalidArgumentException();
 		}
 
@@ -65,15 +61,20 @@ final class DataBlock{
 		// First count the total number of data blocks
 		$totalBlocks  = 0;
 		$ecBlockArray = $ecBlocks->getECBlocks();
+		/** @var \Zxing\Qrcode\Decoder\ECB $ecBlock */
 		foreach($ecBlockArray as $ecBlock){
 			$totalBlocks += $ecBlock->getCount();
 		}
 
 		// Now establish DataBlocks of the appropriate size and number of data codewords
+		/** @var DataBlock[] */
 		$result          = [];//new DataBlock[$totalBlocks];
 		$numResultBlocks = 0;
+
+		/** @var \Zxing\Qrcode\Decoder\ECB $ecBlock */
 		foreach($ecBlockArray as $ecBlock){
 			$ecBlockCount = $ecBlock->getCount();
+
 			for($i = 0; $i < $ecBlockCount; $i++){
 				$numDataCodewords           = $ecBlock->getDataCodewords();
 				$numBlockCodewords          = $ecBlocks->getECCodewordsPerBlock() + $numDataCodewords;
@@ -83,32 +84,40 @@ final class DataBlock{
 
 		// All blocks have the same amount of data, except that the last n
 		// (where n may be 0) have 1 more byte. Figure out where these start.
-		$shorterBlocksTotalCodewords = count($result[0]->codewords);
-		$longerBlocksStartAt         = count($result) - 1;
+		$shorterBlocksTotalCodewords = \count($result[0]->codewords);
+		$longerBlocksStartAt         = \count($result) - 1;
+
 		while($longerBlocksStartAt >= 0){
-			$numCodewords = count($result[$longerBlocksStartAt]->codewords);
+			$numCodewords = \count($result[$longerBlocksStartAt]->codewords);
+
 			if($numCodewords == $shorterBlocksTotalCodewords){
 				break;
 			}
+
 			$longerBlocksStartAt--;
 		}
+
 		$longerBlocksStartAt++;
 
 		$shorterBlocksNumDataCodewords = $shorterBlocksTotalCodewords - $ecBlocks->getECCodewordsPerBlock();
 		// The last elements of result may be 1 element longer;
 		// first fill out as many elements as all of them have
 		$rawCodewordsOffset = 0;
+
 		for($i = 0; $i < $shorterBlocksNumDataCodewords; $i++){
 			for($j = 0; $j < $numResultBlocks; $j++){
 				$result[$j]->codewords[$i] = $rawCodewords[$rawCodewordsOffset++];
 			}
 		}
+
 		// Fill out the last data block in the longer ones
 		for($j = $longerBlocksStartAt; $j < $numResultBlocks; $j++){
 			$result[$j]->codewords[$shorterBlocksNumDataCodewords] = $rawCodewords[$rawCodewordsOffset++];
 		}
+
 		// Now add in error correction blocks
-		$max = count($result[0]->codewords);
+		$max = \count($result[0]->codewords);
+
 		for($i = $shorterBlocksNumDataCodewords; $i < $max; $i++){
 			for($j = 0; $j < $numResultBlocks; $j++){
 				$iOffset                         = $j < $longerBlocksStartAt ? $i : $i + 1;
