@@ -17,8 +17,10 @@
 
 namespace Zxing\Qrcode\Decoder;
 
+use Exception;
 use Zxing\ChecksumException;
 use Zxing\Common\BitMatrix;
+use Zxing\Common\DecoderResult;
 use Zxing\Common\Reedsolomon\GenericGF;
 use Zxing\Common\Reedsolomon\ReedSolomonDecoder;
 use Zxing\Common\Reedsolomon\ReedSolomonException;
@@ -32,14 +34,15 @@ use Zxing\FormatException;
  */
 final class Decoder{
 
-	private $rsDecoder;
+	private ReedSolomonDecoder $rsDecoder;
 
 	public function __construct(){
 		$this->rsDecoder = new ReedSolomonDecoder(GenericGF::$QR_CODE_FIELD_256);
 	}
 
-	public function decode($variable, $hints = null){
-		if(is_array($variable)){
+	public function decode($variable, array $hints = null):DecoderResult{
+
+		if(\is_array($variable)){
 			return $this->decodeImage($variable, $hints);
 		}
 		elseif($variable instanceof BitMatrix){
@@ -48,7 +51,8 @@ final class Decoder{
 		elseif($variable instanceof BitMatrixParser){
 			return $this->decodeParser($variable, $hints);
 		}
-		die('decode error Decoder.php');
+
+		throw new Exception('decode error Decoder.php');
 	}
 
 	/**
@@ -56,15 +60,16 @@ final class Decoder{
 	 * "true" is taken to mean a black module.</p>
 	 *
 	 * @param array $image booleans representing white/black QR Code modules
-	 * @param hints  decoding hints that should be used to influence decoding
+	 * @param array $hints  decoding hints that should be used to influence decoding
 	 *
-	 * @return text and bytes encoded within the QR Code
+	 * @return DecoderResult text and bytes encoded within the QR Code
 	 * @throws FormatException if the QR Code cannot be decoded
 	 * @throws ChecksumException if error correction fails
 	 */
-	public function decodeImage($image, $hints = null){
-		$dimension = count($image);
+	public function decodeImage(array $image, array $hints = null):DecoderResult{
+		$dimension = \count($image);
 		$bits      = new BitMatrix($dimension);
+
 		for($i = 0; $i < $dimension; $i++){
 			for($j = 0; $j < $dimension; $j++){
 				if($image[$i][$j]){
@@ -80,15 +85,15 @@ final class Decoder{
 	 * <p>Decodes a QR Code represented as a {@link BitMatrix}. A 1 or "true" is taken to mean a black module.</p>
 	 *
 	 * @param BitMatrix $bits booleans representing white/black QR Code modules
-	 * @param hints decoding hints that should be used to influence decoding
+	 * @param array $hints decoding hints that should be used to influence decoding
 	 *
-	 * @return text and bytes encoded within the QR Code
+	 * @return DecoderResult text and bytes encoded within the QR Code
 	 * @throws FormatException if the QR Code cannot be decoded
 	 * @throws ChecksumException if error correction fails
 	 */
-	public function decodeBits($bits, $hints = null){
+	public function decodeBits(BitMatrix $bits, array $hints = null):DecoderResult{
 
-// Construct a parser and read version, error-correction level
+		// Construct a parser and read version, error-correction level
 		$parser = new BitMatrixParser($bits);
 		$fe     = null;
 		$ce     = null;
@@ -117,11 +122,11 @@ final class Decoder{
 			$parser->readFormatInformation();
 
 			/*
-			* Since we're here, this means we have successfully detected some kind
-			* of version and format information when mirrored. This is a good sign,
-			* that the QR code may be mirrored, and we should try once more with a
-			* mirrored content.
-			*/
+			 * Since we're here, this means we have successfully detected some kind
+			 * of version and format information when mirrored. This is a good sign,
+			 * that the QR code may be mirrored, and we should try once more with a
+			 * mirrored content.
+			 */
 			// Prepare for a mirrored reading.
 			$parser->mirror();
 
@@ -135,10 +140,10 @@ final class Decoder{
 		}
 		catch(FormatException $e){// catch (FormatException | ChecksumException e) {
 			// Throw the exception from the original reading
-			if($fe != null){
+			if($fe !== null){
 				throw $fe;
 			}
-			if($ce != null){
+			if($ce !== null){
 				throw $ce;
 			}
 			throw $e;
@@ -146,34 +151,36 @@ final class Decoder{
 		}
 	}
 
-	private function decodeParser($parser, $hints = null){
+	private function decodeParser(BitMatrixParser $parser, array $hints = null):DecoderResult{
 		$version = $parser->readVersion();
 		$ecLevel = $parser->readFormatInformation()->getErrorCorrectionLevel();
 
-// Read codewords
+		// Read codewords
 		$codewords = $parser->readCodewords();
-// Separate into data blocks
+		// Separate into data blocks
 		$dataBlocks = DataBlock::getDataBlocks($codewords, $version, $ecLevel);
 
-// Count total number of data bytes
+		// Count total number of data bytes
 		$totalBytes = 0;
 		foreach($dataBlocks as $dataBlock){
 			$totalBytes += $dataBlock->getNumDataCodewords();
 		}
+
 		$resultBytes  = fill_array(0, $totalBytes, 0);
 		$resultOffset = 0;
 
-// Error-correct and copy data blocks together into a stream of bytes
+		// Error-correct and copy data blocks together into a stream of bytes
 		foreach($dataBlocks as $dataBlock){
 			$codewordBytes    = $dataBlock->getCodewords();
 			$numDataCodewords = $dataBlock->getNumDataCodewords();
 			$this->correctErrors($codewordBytes, $numDataCodewords);
+
 			for($i = 0; $i < $numDataCodewords; $i++){
 				$resultBytes[$resultOffset++] = $codewordBytes[$i];
 			}
 		}
 
-// Decode the contents of that stream of bytes
+		// Decode the contents of that stream of bytes
 		return DecodedBitStreamParser::decode($resultBytes, $version, $ecLevel, $hints);
 	}
 
@@ -181,27 +188,30 @@ final class Decoder{
 	 * <p>Given data and error-correction codewords received, possibly corrupted by errors, attempts to
 	 * correct the errors in-place using Reed-Solomon error correction.</p>
 	 *
-	 * @param codewordBytes    data and error correction codewords
-	 * @param numDataCodewords number of codewords that are data bytes
+	 * @param array $codewordBytes    data and error correction codewords
+	 * @param int $numDataCodewords number of codewords that are data bytes
 	 *
 	 * @throws ChecksumException if error correction fails
 	 */
-	private function correctErrors(&$codewordBytes, $numDataCodewords){
-		$numCodewords = count($codewordBytes);
-// First read into an array of ints
+	private function correctErrors(array &$codewordBytes, int $numDataCodewords){
+		$numCodewords = \count($codewordBytes);
+		// First read into an array of ints
 		$codewordsInts = fill_array(0, $numCodewords, 0);
+
 		for($i = 0; $i < $numCodewords; $i++){
 			$codewordsInts[$i] = $codewordBytes[$i] & 0xFF;
 		}
-		$numECCodewords = count($codewordBytes) - $numDataCodewords;
+
+		$numECCodewords = \count($codewordBytes) - $numDataCodewords;
+
 		try{
 			$this->rsDecoder->decode($codewordsInts, $numECCodewords);
 		}
 		catch(ReedSolomonException $ignored){
-			throw ChecksumException::getChecksumInstance();
+			throw new ChecksumException();
 		}
-// Copy back into array of bytes -- only need to worry about the bytes that were data
-// We don't care about errors in the error-correction codewords
+		// Copy back into array of bytes -- only need to worry about the bytes that were data
+		// We don't care about errors in the error-correction codewords
 		for($i = 0; $i < $numDataCodewords; $i++){
 			$codewordBytes[$i] = $codewordsInts[$i];
 		}
