@@ -17,6 +17,7 @@
 
 namespace Zxing\Qrcode\Detector;
 
+use Zxing\Common\BitMatrix;
 use Zxing\NotFoundException;
 
 /**
@@ -35,54 +36,52 @@ use Zxing\NotFoundException;
  */
 final class AlignmentPatternFinder{
 
-	private $image;
-	private $possibleCenters;
-	private $startX;
-	private $startY;
-	private $width;
-	private $height;
-	private $moduleSize;
-	private $crossCheckStateCount;
-	private $resultPointCallback;
+	private BitMatrix $image;
+	private int $startX;
+	private int $startY;
+	private int $width;
+	private int $height;
+	private float $moduleSize;
+	/** @var \Zxing\Qrcode\Detector\AlignmentPattern[] */
+	private array $possibleCenters;
+	private array $crossCheckStateCount;
 
 	/**
 	 * <p>Creates a finder that will look in a portion of the whole image.</p>
 	 *
-	 * @param image      image to search
-	 * @param startX     left column from which to start searching
-	 * @param startY     top row from which to start searching
-	 * @param width      width of region to search
-	 * @param height     height of region to search
-	 * @param moduleSize estimated module size so far
+	 * @param \Zxing\Common\BitMatrix $image      image to search
+	 * @param int                     $startX     left column from which to start searching
+	 * @param int                     $startY     top row from which to start searching
+	 * @param int                     $width      width of region to search
+	 * @param int                     $height     height of region to search
+	 * @param float                   $moduleSize estimated module size so far
 	 */
 	public function __construct(
-		$image,
-		$startX,
-		$startY,
-		$width,
-		$height,
-		$moduleSize,
-		$resultPointCallback
+		BitMatrix $image,
+		int $startX,
+		int $startY,
+		int $width,
+		int $height,
+		float $moduleSize
 	){
 		$this->image                = $image;
-		$this->possibleCenters      = [];
 		$this->startX               = $startX;
 		$this->startY               = $startY;
 		$this->width                = $width;
 		$this->height               = $height;
 		$this->moduleSize           = $moduleSize;
+		$this->possibleCenters      = [];
 		$this->crossCheckStateCount = [];
-		$this->resultPointCallback  = $resultPointCallback;
 	}
 
 	/**
 	 * <p>This method attempts to find the bottom-right alignment pattern in the image. It is a bit messy since
 	 * it's pretty performance-critical and so is written to be fast foremost.</p>
 	 *
-	 * @return {@link AlignmentPattern} if found
+	 * @return \Zxing\Qrcode\Detector\AlignmentPattern if found
 	 * @throws NotFoundException if not found
 	 */
-	public function find(){
+	public function find():AlignmentPattern{
 		$startX  = $this->startX;
 		$height  = $this->height;
 		$maxJ    = $startX + $this->width;
@@ -90,10 +89,10 @@ final class AlignmentPatternFinder{
 		// We are looking for black/white/black modules in 1:1:1 ratio;
 		// this tracks the number of black/white/black modules seen so far
 		$stateCount = [];
+
 		for($iGen = 0; $iGen < $height; $iGen++){
 			// Search from middle outwards
-			$i             = $middleI + (($iGen & 0x01) == 0 ? ($iGen + 1) / 2 : -(($iGen + 1) / 2));
-			$i             = (int)($i);
+			$i             = (int)($middleI + (($iGen & 0x01) === 0 ? ($iGen + 1) / 2 : -(($iGen + 1) / 2)));
 			$stateCount[0] = 0;
 			$stateCount[1] = 0;
 			$stateCount[2] = 0;
@@ -104,21 +103,28 @@ final class AlignmentPatternFinder{
 			while($j < $maxJ && !$this->image->get($j, $i)){
 				$j++;
 			}
+
 			$currentState = 0;
+
 			while($j < $maxJ){
+
 				if($this->image->get($j, $i)){
 					// Black pixel
-					if($currentState == 1){ // Counting black pixels
+					if($currentState === 1){ // Counting black pixels
 						$stateCount[$currentState]++;
 					}
 					else{ // Counting white pixels
-						if($currentState == 2){ // A winner?
+
+						if($currentState === 2){ // A winner?
+
 							if($this->foundPatternCross($stateCount)){ // Yes
 								$confirmed = $this->handlePossibleCenter($stateCount, $i, $j);
-								if($confirmed != null){
+
+								if($confirmed !== null){
 									return $confirmed;
 								}
 							}
+
 							$stateCount[0] = $stateCount[2];
 							$stateCount[1] = 1;
 							$stateCount[2] = 0;
@@ -130,16 +136,21 @@ final class AlignmentPatternFinder{
 					}
 				}
 				else{ // White pixel
-					if($currentState == 1){ // Counting black pixels
+
+					if($currentState === 1){ // Counting black pixels
 						$currentState++;
 					}
+
 					$stateCount[$currentState]++;
 				}
+
 				$j++;
 			}
+
 			if($this->foundPatternCross($stateCount)){
 				$confirmed = $this->handlePossibleCenter($stateCount, $i, $maxJ);
-				if($confirmed != null){
+
+				if($confirmed !== null){
 					return $confirmed;
 				}
 			}
@@ -148,24 +159,25 @@ final class AlignmentPatternFinder{
 
 		// Hmm, nothing we saw was observed and confirmed twice. If we had
 		// any guess at all, return it.
-		if(count($this->possibleCenters)){
+		if(\count($this->possibleCenters)){
 			return $this->possibleCenters[0];
 		}
 
-		throw  NotFoundException::getNotFoundInstance();
+		throw new NotFoundException();
 	}
 
 	/**
-	 * @param stateCount count of black/white/black pixels just read
+	 * @param int[] $stateCount count of black/white/black pixels just read
 	 *
-	 * @return true iff the proportions of the counts is close enough to the 1/1/1 ratios
+	 * @return bool true if the proportions of the counts is close enough to the 1/1/1 ratios
 	 *         used by alignment patterns to be considered a match
 	 */
-	private function foundPatternCross($stateCount){
+	private function foundPatternCross(array $stateCount):bool{
 		$moduleSize  = $this->moduleSize;
 		$maxVariance = $moduleSize / 2.0;
+
 		for($i = 0; $i < 3; $i++){
-			if(abs($moduleSize - $stateCount[$i]) >= $maxVariance){
+			if(\abs($moduleSize - $stateCount[$i]) >= $maxVariance){
 				return false;
 			}
 		}
@@ -179,30 +191,30 @@ final class AlignmentPatternFinder{
 	 * found on a previous horizontal scan. If so, we consider it confirmed and conclude we have
 	 * found the alignment pattern.</p>
 	 *
-	 * @param stateCount reading state module counts from horizontal scan
-	 * @param i          row where alignment pattern may be found
-	 * @param j          end of possible alignment pattern in row
+	 * @param int[] $stateCount reading state module counts from horizontal scan
+	 * @param int   $i          row where alignment pattern may be found
+	 * @param int   $j          end of possible alignment pattern in row
 	 *
-	 * @return {@link AlignmentPattern} if we have found the same pattern twice, or null if not
+	 * @return \Zxing\Qrcode\Detector\AlignmentPattern|null if we have found the same pattern twice, or null if not
 	 */
-	private function handlePossibleCenter($stateCount, $i, $j){
+	private function handlePossibleCenter(array $stateCount, int $i, int $j):?AlignmentPattern{
 		$stateCountTotal = $stateCount[0] + $stateCount[1] + $stateCount[2];
 		$centerJ         = $this->centerFromEnd($stateCount, $j);
 		$centerI         = $this->crossCheckVertical($i, (int)$centerJ, 2 * $stateCount[1], $stateCountTotal);
+
 		if(!is_nan($centerI)){
 			$estimatedModuleSize = (float)($stateCount[0] + $stateCount[1] + $stateCount[2]) / 3.0;
+
 			foreach($this->possibleCenters as $center){
 				// Look for about the same center and module size:
 				if($center->aboutEquals($estimatedModuleSize, $centerI, $centerJ)){
 					return $center->combineEstimate($centerI, $centerJ, $estimatedModuleSize);
 				}
 			}
+
 			// Hadn't found this before; save it
 			$point                   = new AlignmentPattern($centerJ, $centerI, $estimatedModuleSize);
 			$this->possibleCenters[] = $point;
-			if($this->resultPointCallback != null){
-				$this->resultPointCallback->foundPossibleResultPoint($point);
-			}
 		}
 
 		return null;
@@ -211,9 +223,14 @@ final class AlignmentPatternFinder{
 	/**
 	 * Given a count of black/white/black pixels just seen and an end position,
 	 * figures the location of the center of this black/white/black run.
+	 *
+	 * @param int[] $stateCount
+	 * @param int   $end
+	 *
+	 * @return float
 	 */
-	private static function centerFromEnd($stateCount, $end){
-		return (float)($end - $stateCount[2]) - $stateCount[1] / 2.0;
+	private function centerFromEnd(array $stateCount, int $end):float{
+		return (float)(($end - $stateCount[2]) - $stateCount[1] / 2.0);
 	}
 
 	/**
@@ -221,20 +238,16 @@ final class AlignmentPatternFinder{
 	 * "cross-checks" by scanning down vertically through the center of the possible
 	 * alignment pattern to see if the same proportion is detected.</p>
 	 *
-	 * @param startI   row where an alignment pattern was detected
-	 * @param centerJ  center of the section that appears to cross an alignment pattern
-	 * @param maxCount maximum reasonable number of modules that should be
-	 *                 observed in any reading state, based on the results of the horizontal scan
+	 * @param int $startI   row where an alignment pattern was detected
+	 * @param int $centerJ  center of the section that appears to cross an alignment pattern
+	 * @param int $maxCount maximum reasonable number of modules that should be
+	 *                      observed in any reading state, based on the results of the horizontal scan
+	 * @param int $originalStateCountTotal
 	 *
-	 * @return vertical center of alignment pattern, or {@link Float#NaN} if not found
+	 * @return float vertical center of alignment pattern, or NAN if not found
 	 */
-	private function crossCheckVertical(
-		$startI, $centerJ, $maxCount,
-		$originalStateCountTotal
-	){
-		$image = $this->image;
-
-		$maxI          = $image->getHeight();
+	private function crossCheckVertical(int $startI, int $centerJ, int $maxCount, int $originalStateCountTotal):float{
+		$maxI          = $this->image->getHeight();
 		$stateCount    = $this->crossCheckStateCount;
 		$stateCount[0] = 0;
 		$stateCount[1] = 0;
@@ -242,44 +255,48 @@ final class AlignmentPatternFinder{
 
 		// Start counting up from center
 		$i = $startI;
-		while($i >= 0 && $image->get($centerJ, $i) && $stateCount[1] <= $maxCount){
+		while($i >= 0 && $this->image->get($centerJ, $i) && $stateCount[1] <= $maxCount){
 			$stateCount[1]++;
 			$i--;
 		}
 		// If already too many modules in this state or ran off the edge:
 		if($i < 0 || $stateCount[1] > $maxCount){
-			return NAN;
+			return \NAN;
 		}
-		while($i >= 0 && !$image->get($centerJ, $i) && $stateCount[0] <= $maxCount){
+
+		while($i >= 0 && !$this->image->get($centerJ, $i) && $stateCount[0] <= $maxCount){
 			$stateCount[0]++;
 			$i--;
 		}
+
 		if($stateCount[0] > $maxCount){
-			return NAN;
+			return \NAN;
 		}
 
 		// Now also count down from center
 		$i = $startI + 1;
-		while($i < $maxI && $image->get($centerJ, $i) && $stateCount[1] <= $maxCount){
+		while($i < $maxI && $this->image->get($centerJ, $i) && $stateCount[1] <= $maxCount){
 			$stateCount[1]++;
 			$i++;
 		}
+
 		if($i == $maxI || $stateCount[1] > $maxCount){
-			return NAN;
+			return \NAN;
 		}
-		while($i < $maxI && !$image->get($centerJ, $i) && $stateCount[2] <= $maxCount){
+
+		while($i < $maxI && !$this->image->get($centerJ, $i) && $stateCount[2] <= $maxCount){
 			$stateCount[2]++;
 			$i++;
 		}
+
 		if($stateCount[2] > $maxCount){
-			return NAN;
+			return \NAN;
 		}
 
-		$stateCountTotal = $stateCount[0] + $stateCount[1] + $stateCount[2];
-		if(5 * abs($stateCountTotal - $originalStateCountTotal) >= 2 * $originalStateCountTotal){
-			return NAN;
+		if(5 * \abs(($stateCount[0] + $stateCount[1] + $stateCount[2]) - $originalStateCountTotal) >= 2 * $originalStateCountTotal){
+			return \NAN;
 		}
 
-		return $this->foundPatternCross($stateCount) ? $this->centerFromEnd($stateCount, $i) : NAN;
+		return $this->foundPatternCross($stateCount) ? $this->centerFromEnd($stateCount, $i) : \NAN;
 	}
 }
