@@ -143,7 +143,7 @@ class MonochromeRectangleDetector{
 			}
 			if($range == null){
 				if($lastRange == null){
-					throw NotFoundException::getNotFoundInstance();
+					throw new NotFoundException();
 				}
 				// lastRange was found
 				if($deltaX == 0){
@@ -176,7 +176,7 @@ class MonochromeRectangleDetector{
 			}
 			$lastRange = $range;
 		}
-		throw NotFoundException::getNotFoundInstance();
+		throw new NotFoundException();
 	}
 
 	/**
