@@ -18,6 +18,7 @@
 namespace Zxing\Common;
 
 use Zxing\Binarizer;
+use Zxing\LuminanceSource;
 use Zxing\NotFoundException;
 
 /**
@@ -33,32 +34,27 @@ use Zxing\NotFoundException;
  */
 class GlobalHistogramBinarizer extends Binarizer{
 
-	private static $LUMINANCE_BITS    = 5;
-	private static $LUMINANCE_SHIFT   = 3;
-	private static $LUMINANCE_BUCKETS = 32;
+	private const LUMINANCE_BITS    = 5;
+	private const LUMINANCE_SHIFT   = 3;
+	private const LUMINANCE_BUCKETS = 32;
 
-	private static $EMPTY = [];
+	private array $luminances;
+	private array $buckets;
 
-	private $luminances = [];
-	private $buckets    = [];
-	private $source     = [];
-
-	public function __construct($source){
-		self::$LUMINANCE_SHIFT   = 8 - self::$LUMINANCE_BITS;
-		self::$LUMINANCE_BUCKETS = 1 << self::$LUMINANCE_BITS;
-
+	public function __construct(LuminanceSource $source){
 		parent::__construct($source);
 
-		$this->luminances = self::$EMPTY;
-		$this->buckets    = fill_array(0, self::$LUMINANCE_BUCKETS, 0);
+		$this->luminances = [];
+		$this->buckets    = fill_array(0, self::LUMINANCE_BUCKETS, 0);
 		$this->source     = $source;
 	}
 
 	// Applies simple sharpening to the row data to improve performance of the 1D Readers.
-	public function getBlackRow($y, $row = null){
+	public function getBlackRow(int $y, BitArray $row = null):BitArray{
 		$this->source = $this->getLuminanceSource();
 		$width        = $this->source->getWidth();
-		if($row == null || $row->getSize() < $width){
+
+		if($row === null || $row->getSize() < $width){
 			$row = new BitArray($width);
 		}
 		else{
@@ -68,21 +64,25 @@ class GlobalHistogramBinarizer extends Binarizer{
 		$this->initArrays($width);
 		$localLuminances = $this->source->getRow($y, $this->luminances);
 		$localBuckets    = $this->buckets;
+
 		for($x = 0; $x < $width; $x++){
 			$pixel = $localLuminances[$x] & 0xff;
-			$localBuckets[$pixel >> self::$LUMINANCE_SHIFT]++;
+			$localBuckets[$pixel >> self::LUMINANCE_SHIFT]++;
 		}
-		$blackPoint = self::estimateBlackPoint($localBuckets);
 
+		$blackPoint = $this->estimateBlackPoint($localBuckets);
 		$left   = $localLuminances[0] & 0xff;
 		$center = $localLuminances[1] & 0xff;
+
 		for($x = 1; $x < $width - 1; $x++){
 			$right = $localLuminances[$x + 1] & 0xff;
 			// A simple -1 4 -1 box filter with a weight of 2.
 			$luminance = (($center * 4) - $left - $right) / 2;
+
 			if($luminance < $blackPoint){
 				$row->set($x);
 			}
+
 			$left   = $center;
 			$center = $right;
 		}
@@ -92,25 +92,31 @@ class GlobalHistogramBinarizer extends Binarizer{
 
 	// Does not sharpen the data, as this call is intended to only be used by 2D Readers.
 	private function initArrays($luminanceSize){
-		if(count($this->luminances) < $luminanceSize){
+
+		if(\count($this->luminances) < $luminanceSize){
 			$this->luminances = [];
 		}
-		for($x = 0; $x < self::$LUMINANCE_BUCKETS; $x++){
+
+		for($x = 0; $x < self::LUMINANCE_BUCKETS; $x++){
 			$this->buckets[$x] = 0;
 		}
+
 	}
 
-	private static function estimateBlackPoint($buckets){
+	private function estimateBlackPoint(array $buckets):int{
 		// Find the tallest peak in the histogram.
-		$numBuckets     = count($buckets);
+		$numBuckets     = \count($buckets);
 		$maxBucketCount = 0;
 		$firstPeak      = 0;
 		$firstPeakSize  = 0;
+
 		for($x = 0; $x < $numBuckets; $x++){
+
 			if($buckets[$x] > $firstPeakSize){
 				$firstPeak     = $x;
 				$firstPeakSize = $buckets[$x];
 			}
+
 			if($buckets[$x] > $maxBucketCount){
 				$maxBucketCount = $buckets[$x];
 			}
@@ -119,10 +125,12 @@ class GlobalHistogramBinarizer extends Binarizer{
 		// Find the second-tallest peak which is somewhat far from the tallest peak.
 		$secondPeak      = 0;
 		$secondPeakScore = 0;
+
 		for($x = 0; $x < $numBuckets; $x++){
 			$distanceToBiggest = $x - $firstPeak;
 			// Encourage more distant second peaks by multiplying by square of distance.
 			$score = $buckets[$x] * $distanceToBiggest * $distanceToBiggest;
+
 			if($score > $secondPeakScore){
 				$secondPeak      = $x;
 				$secondPeakScore = $score;
@@ -139,25 +147,27 @@ class GlobalHistogramBinarizer extends Binarizer{
 		// If there is too little contrast in the image to pick a meaningful black point, throw rather
 		// than waste time trying to decode the image, and risk false positives.
 		if($secondPeak - $firstPeak <= $numBuckets / 16){
-			throw NotFoundException::getNotFoundInstance();
+			throw new NotFoundException();
 		}
 
 		// Find a valley between them that is low and closer to the white peak.
 		$bestValley      = $secondPeak - 1;
 		$bestValleyScore = -1;
+
 		for($x = $secondPeak - 1; $x > $firstPeak; $x--){
 			$fromFirst = $x - $firstPeak;
 			$score     = $fromFirst * $fromFirst * ($secondPeak - $x) * ($maxBucketCount - $buckets[$x]);
+
 			if($score > $bestValleyScore){
 				$bestValley      = $x;
 				$bestValleyScore = $score;
 			}
 		}
 
-		return ($bestValley << self::$LUMINANCE_SHIFT);
+		return $bestValley << self::LUMINANCE_SHIFT;
 	}
 
-	public function getBlackMatrix(){
+	public function getBlackMatrix():BitMatrix{
 		$source = $this->getLuminanceSource();
 		$width  = $source->getWidth();
 		$height = $source->getHeight();
@@ -167,25 +177,31 @@ class GlobalHistogramBinarizer extends Binarizer{
 		// more robust on the blackbox tests than sampling a diagonal as we used to do.
 		$this->initArrays($width);
 		$localBuckets = $this->buckets;
+
 		for($y = 1; $y < 5; $y++){
 			$row             = (int)($height * $y / 5);
 			$localLuminances = $source->getRow($row, $this->luminances);
 			$right           = (int)(($width * 4) / 5);
+
 			for($x = (int)($width / 5); $x < $right; $x++){
-				$pixel = ($localLuminances[(int)($x)] & 0xff);
-				$localBuckets[($pixel >> self::$LUMINANCE_SHIFT)]++;
+				$pixel = $localLuminances[(int)$x] & 0xff;
+				$localBuckets[$pixel >> self::LUMINANCE_SHIFT]++;
 			}
 		}
-		$blackPoint = self::estimateBlackPoint($localBuckets);
+
+		$blackPoint = $this->estimateBlackPoint($localBuckets);
 
 		// We delay reading the entire image luminance until the black point estimation succeeds.
 		// Although we end up reading four rows twice, it is consistent with our motto of
 		// "fail quickly" which is necessary for continuous scanning.
 		$localLuminances = $source->getMatrix();
+
 		for($y = 0; $y < $height; $y++){
 			$offset = $y * $width;
+
 			for($x = 0; $x < $width; $x++){
 				$pixel = (int)($localLuminances[$offset + $x] & 0xff);
+
 				if($pixel < $blackPoint){
 					$matrix->set($x, $y);
 				}
@@ -195,7 +211,7 @@ class GlobalHistogramBinarizer extends Binarizer{
 		return $matrix;
 	}
 
-	public function createBinarizer($source){
+	public function createBinarizer(LuminanceSource $source):Binarizer{
 		return new GlobalHistogramBinarizer($source);
 	}
 }
