@@ -34,16 +34,16 @@ final class GenericGFPoly{
 	private $coefficients;
 
 	/**
-	 * @param field        the {@link GenericGF} instance representing the field to use
+	 * @param GenericGF $field        the {@link GenericGF} instance representing the field to use
 	 *                     to perform computations
-	 * @param coefficients array coefficients as ints representing elements of GF(size), arranged
+	 * @param array $coefficients array coefficients as ints representing elements of GF(size), arranged
 	 *                     from most significant (highest-power term) coefficient to least significant
 	 *
 	 * @throws InvalidArgumentException if argument is null or empty,
 	 * or if leading coefficient is 0 and this is not a
 	 * constant polynomial (that is, it is not the monomial "0")
 	 */
-	public function __construct($field, $coefficients){
+	public function __construct(GenericGF $field, array $coefficients){
 		if(count($coefficients) == 0){
 			throw new InvalidArgumentException();
 		}
@@ -242,7 +242,7 @@ final class GenericGFPoly{
 		return new GenericGFPoly($this->field, $sumDiff);
 	}
 
-	//@Override
+
 
 	public function toString(){
 		$result = '';
