@@ -23,15 +23,4 @@ namespace Zxing;
  *
  * @author Sean Owen
  */
-final class NotFoundException extends ReaderException{
-
-	private static $instance;
-
-	public static function getNotFoundInstance(){
-		if(!self::$instance){
-			self::$instance = new NotFoundException();
-		}
-
-		return self::$instance;
-	}
-}
+class NotFoundException extends ReaderException{}
