@@ -17,7 +17,9 @@
 
 namespace Zxing\Qrcode\Decoder;
 
+use Closure;
 use InvalidArgumentException;
+use Zxing\Common\BitMatrix;
 
 /**
  * <p>Encapsulates data masks for the data bits in a QR code, per ISO 18004:2006 6.8. Implementations
@@ -30,154 +32,51 @@ use InvalidArgumentException;
  *
  * @author Sean Owen
  */
-abstract class DataMask{
-
-	/**
-	 * See ISO 18004:2006 6.8.1
-	 */
-	private static $DATA_MASKS = [];
-
-	public function __construct(){
-
-	}
-
-	public static function Init(){
-		self::$DATA_MASKS = [
-			new DataMask000(),
-			new DataMask001(),
-			new DataMask010(),
-			new DataMask011(),
-			new DataMask100(),
-			new DataMask101(),
-			new DataMask110(),
-			new DataMask111(),
-		];
-	}
-
-	/**
-	 * @param reference a value between 0 and 7 indicating one of the eight possible
-	 *                  data mask patterns a QR Code may use
-	 *
-	 * @return DataMask encapsulating the data mask pattern
-	 */
-	public static function forReference($reference){
-		if($reference < 0 || $reference > 7){
-			throw new InvalidArgumentException();
-		}
-
-		return self::$DATA_MASKS[$reference];
-	}
+class DataMask{
 
 	/**
 	 * <p>Implementations of this method reverse the data masking process applied to a QR Code and
 	 * make its bits ready to read.</p>
 	 *
-	 * @param bits      representation of QR Code bits
-	 * @param dimension dimension of QR Code, represented by bits, being unmasked
+	 * @param \Zxing\Common\BitMatrix $bits
+	 * @param int                     $dimension
+	 * @param int                     $maskPattern
+	 *
+	 * @return \Zxing\Common\BitMatrix
 	 */
-	final public function unmaskBitMatrix($bits, $dimension){
+	public static function unmaskBitMatrix(BitMatrix $bits, int $dimension, int $maskPattern):BitMatrix{
+		$mask = self::getMask($maskPattern);
+
 		for($i = 0; $i < $dimension; $i++){
 			for($j = 0; $j < $dimension; $j++){
-				if($this->isMasked($i, $j)){
+				if($mask($i, $j) === 0){
 					$bits->flip($j, $i);
 				}
 			}
 		}
+
+		return $bits;
 	}
 
-	abstract public function isMasked($i, $j);
-}
+	/**
+	 * @param int $maskPattern a value between 0 and 7 indicating one of the eight possible
+	 *                 data mask patterns a QR Code may use
+	 *
+	 * @return \Closure
+	 */
+	protected static function getMask(int $maskPattern):Closure{
 
-DataMask::Init();
+		return [
+			0b000 => fn($i, $j):int => ($i + $j) % 2,
+			0b001 => fn($i, $j):int => $i % 2,
+			0b010 => fn($i, $j):int => $j % 3,
+			0b011 => fn($i, $j):int => ($i + $j) % 3,
+			0b100 => fn($i, $j):int => ((int)($i / 2) + (int)($j / 3)) % 2,
+			0b101 => fn($i, $j):int => (($i * $j) % 2) + (($i * $j) % 3),
+			0b110 => fn($i, $j):int => ((($i * $j) % 2) + (($i * $j) % 3)) % 2,
+			0b111 => fn($i, $j):int => ((($i * $j) % 3) + (($i + $j) % 2)) % 2,
+		][$maskPattern];
 
-/**
- * 000: mask bits for which (x + y) mod 2 == 0
- */
-final class DataMask000 extends DataMask{
-
-	// @Override
-	public function isMasked($i, $j){
-		return (($i + $j) & 0x01) == 0;
 	}
-}
 
-/**
- * 001: mask bits for which x mod 2 == 0
- */
-final class DataMask001 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		return ($i & 0x01) == 0;
-	}
-}
-
-/**
- * 010: mask bits for which y mod 3 == 0
- */
-final class DataMask010 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		return $j % 3 == 0;
-	}
-}
-
-/**
- * 011: mask bits for which (x + y) mod 3 == 0
- */
-final class DataMask011 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		return ($i + $j) % 3 == 0;
-	}
-}
-
-/**
- * 100: mask bits for which (x/2 + y/3) mod 2 == 0
- */
-final class DataMask100 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		return (int)(((int)($i / 2) + (int)($j / 3)) & 0x01) == 0;
-	}
-}
-
-/**
- * 101: mask bits for which xy mod 2 + xy mod 3 == 0
- */
-final class DataMask101 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		$temp = $i * $j;
-
-		return ($temp & 0x01) + ($temp % 3) == 0;
-	}
-}
-
-/**
- * 110: mask bits for which (xy mod 2 + xy mod 3) mod 2 == 0
- */
-final class DataMask110 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		$temp = $i * $j;
-
-		return ((($temp & 0x01) + ($temp % 3)) & 0x01) == 0;
-	}
-}
-
-/**
- * 111: mask bits for which ((x+y)mod 2 + xy mod 3) mod 2 == 0
- */
-final class DataMask111 extends DataMask{
-
-	//@Override
-	public function isMasked($i, $j){
-		return (((($i + $j) & 0x01) + (($i * $j) % 3)) & 0x01) == 0;
-	}
 }
